Shifting thru specified SQL columns.
This project snippet shifts thru an undetermined amount of columns named "Ing n" where "Ing" stands for Ingredient and "n" is the number corresponding to the specific column.
The code is implemented when a user opens the application. As it checks if a recipe has been deleted.
If a row in the recipe table, that contains an Ingredient only that the recipe uses, has been dropped, then it deletes the Ingredient is dropped from the Item table.
If a recipe has been added containing a new Ingredient then that ingredient is added to the Item table. 
This was implemented since the admins had to add recipes directly into the database, and not through an interactive system, where other code could have added the item directly when a new recipe was made.